WatchNet/event_router/requirements.txt
fastapi[all]~=0.115.4
uvicorn
pydantic~=2.9.2
kafka-python
python-dotenv
elasticsearch
sqlalchemy